Correctly scale axis around mouse
The calculation of new axis scale min and max values when
`IsZoomOnMouseCenter = true` causes the scale to center on the mouse over
value. This causes erratic zooming behavior when using the mouse wheel
as the mouse over value continuously changes.

Instead of centering on the mouse over value, we need to center the zoom
around the mouse over value.