Remove demo and test applications
By removing the demo and test application this repository will be a
library only repository, and allow this repository to focus on moving
the library forward to support net standard, other graphics libraries and
UI frameworks. One side-effect of this will be less evident usage of the
library and will require a more diligent analysis and discipline to not
cause inadvertant breaking changes.