Encapsulate usage of linearized min/max property
To remove the InternalsVisibleTo attribute, the usage of the _minLinearized
and _maxLinerized properties have to be modifed.

To keep the public API somewhat agnostic of the linearity of the different
scales and avoid a MinLinearized/MaxLinearized property being exposed, Any
writes of the internal linear properties have been moved into the new
Scale members: Zoom, Pan, and Scroll. The reading of the linear values in
the control has been converted to use the Scale's Min/Max values along
with the Scale's Linearize methods.